% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{shinyhub_test_slug}
\alias{shinyhub_test_slug}
\title{Default app slug for test tokens.}
\usage{
shinyhub_test_slug()
}
\value{
The slug test tokens are minted for, as a string.
}
\description{
Default app slug for test tokens.
}
\examples{
shinyhub_test_slug()
}
